refactor(project): fold system seam into python
